requirements.txt
Authlib==1.2.0
Flask==2.2.2
Werkzeug==2.2.2
PyJWT==2.7.0
requests==2.28.2
python-dotenv==0.21.0
